backend/requirements.txt
anthropic>=0.40.0
openai>=1.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
mempalace>=0.1.0
